backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sentence-transformers
faiss-cpu
numpy
transformers
torch
sqlparse
